feat: refactor component name (#354)
