+++ app/components/Detail/components/ViewComponentAtlases/viewComponentAtlases.tsx
@@ -10,8 +10,8 @@ import {
   ButtonLink,
   BUTTON_COLOR,
 } from "../../../common/Button/components/ButtonLink/buttonLink";
-import { TableEmpty } from "../../../Table/components/TableEmpty/tableEmpty";
 import { Paper } from "../../../Table/components/TablePaper/tablePaper.styles";
+import { TablePlaceholder } from "../../../Table/components/TablePlaceholder/tablePlaceholder";
 import { Toolbar } from "../../../Table/components/TableToolbar/tableToolbar.styles";
 import { Table } from "../../../Table/table.styles";
 import { RequestAccess } from "./components/RequestAccess/requestAccess";
@@ -52,7 +52,7 @@ export const ViewComponentAtlases = ({
             items={componentAtlases}
           />
         )}
-        <TableEmpty
+        <TablePlaceholder
           canEdit={canEdit}
           message="No integration objects"
           rowCount={componentAtlases.length}
