+++ app/components/Detail/components/ViewSourceDatasets/viewSourceDatasets.tsx
@@ -4,8 +4,8 @@ import { HCAAtlasTrackerSourceDataset } from "../../../../apis/catalog/hca-atlas
 import { PathParameter } from "../../../../common/entities";
 import { FormManager as FormManagerProps } from "../../../../hooks/useFormManager/common/entities";
 import { getAtlasSourceDatasetsTableColumns } from "../../../../viewModelBuilders/catalog/hca-atlas-tracker/common/viewModelBuilders";
-import { TableEmpty } from "../../../Table/components/TableEmpty/tableEmpty";
 import { Paper } from "../../../Table/components/TablePaper/tablePaper.styles";
+import { TablePlaceholder } from "../../../Table/components/TablePlaceholder/tablePlaceholder";
 import { Toolbar } from "../../../Table/components/TableToolbar/tableToolbar.styles";
 import { Table } from "../../../Table/table.styles";
 import { AddSourceDataset } from "../AddSourceDataset/addSourceDataset";
@@ -43,7 +43,7 @@ export const ViewSourceDatasets = ({
             items={sourceDatasets.sort(sortSourceDataset)}
           />
         )}
-        <TableEmpty
+        <TablePlaceholder
           canEdit={canEdit}
           message="No source datasets"
           rowCount={sourceDatasets.length}
