+++ app/components/Table/components/TablePlaceholder/tablePlaceholder.styles.ts
